Deploy router from disk on btrfs subvol (#873)
* Use local ros instead of ovs


* Deploy routeros from local disk on btrfs subvolume

Handle stop/start backup/restore

+++ libs/agent-scripts/vfw/vfs_applyconfig_routeros.py
@@ -11,16 +11,19 @@ author = "hendrik@incubaid.com"
 license = "bsd"
 version = "1.0"
 roles = []
-async = True 
+async = True
+
 
 def action(name, fwobject):
     host = fwobject['host']
     username = fwobject['username']
     password = fwobject['password']
 
+    if not j.system.net.waitConnectionTest(host, 8728, timeout=30):
+        raise RuntimeError("Failed to get connection to api")
     ro = j.clients.routeros.get(host, username, password)
     ro.deletePortForwardRules(tags='cloudbroker')
-    for rule in fwobject['tcpForwardRules']:  
+    for rule in fwobject['tcpForwardRules']:
         protocol = rule.get('protocol', 'tcp')
         ro.addPortForwardRule(rule['fromAddr'], rule['fromPort'], rule['toAddr'], rule['toPort'], tags='cloudbroker', protocol=protocol)
     leases = fwobject.get('leases')
